# src/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Outpoint": {
				"type": "string"
			},
			"File": {
				"properties": {
					"hash": {
						"type": "string"
					},
					"size": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Inscription": {
				"properties": {
					"json": {},
					"text": {
						"type": "string"
					},
					"words": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"file": {
						"$ref": "#/components/schemas/File"
					},
					"parent": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Sigma": {
				"properties": {
					"algorithm": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"signature": {
						"type": "string"
					},
					"vin": {
						"type": "number",
						"format": "double"
					},
					"valid": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Bsv20Status": {
				"enum": [
					-1,
					0,
					1
				],
				"type": "number"
			},
			"TxoData": {
				"properties": {
					"types": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"insc": {
						"$ref": "#/components/schemas/Inscription"
					},
					"map": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"b": {
						"$ref": "#/components/schemas/File"
					},
					"sigma": {
						"items": {
							"$ref": "#/components/schemas/Sigma"
						},
						"type": "array"
					},
					"list": {
						"properties": {
							"sale": {
								"type": "boolean"
							},
							"payout": {
								"type": "string"
							},
							"price": {
								"type": "number",
								"format": "double"
							}
						},
						"type": "object"
					},
					"bsv20": {
						"properties": {
							"implied": {
								"type": "boolean"
							},
							"status": {
								"$ref": "#/components/schemas/Bsv20Status"
							},
							"amt": {
								"type": "string"
							},
							"sym": {
								"type": "string"
							},
							"tick": {
								"type": "string"
							},
							"op": {
								"type": "string"
							},
							"p": {
								"type": "string"
							},
							"id": {
								"$ref": "#/components/schemas/Outpoint"
							}
						},
						"type": "object"
					},
					"lock": {
						"properties": {
							"until": {
								"type": "number",
								"format": "double"
							},
							"address": {
								"type": "string"
							}
						},
						"required": [
							"until",
							"address"
						],
						"type": "object"
					},
					"sigil": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"opns": {
						"properties": {
							"status": {
								"type": "number",
								"format": "double"
							},
							"domain": {
								"type": "string"
							},
							"genesis": {
								"type": "string"
							}
						},
						"type": "object"
					},
					"opnsMine": {
						"properties": {
							"pow": {
								"type": "string"
							},
							"status": {
								"type": "number",
								"format": "double"
							},
							"domain": {
								"type": "string"
							},
							"genesis": {
								"type": "string"
							}
						},
						"type": "object"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Origin": {
				"properties": {
					"outpoint": {
						"$ref": "#/components/schemas/Outpoint"
					},
					"data": {
						"$ref": "#/components/schemas/TxoData"
					},
					"num": {
						"type": "number",
						"format": "double"
					},
					"map": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"inum": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"outpoint"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Txo": {
				"properties": {
					"txid": {
						"type": "string",
						"default": ""
					},
					"vout": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"outpoint": {
						"$ref": "#/components/schemas/Outpoint"
					},
					"satoshis": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"accSats": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"owner": {
						"type": "string"
					},
					"script": {
						"type": "string"
					},
					"spend": {
						"type": "string"
					},
					"spend_height": {
						"type": "number",
						"format": "double"
					},
					"spend_idx": {
						"type": "number",
						"format": "double"
					},
					"origin": {
						"$ref": "#/components/schemas/Origin"
					},
					"height": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"idx": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"data": {
						"$ref": "#/components/schemas/TxoData"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"SortDirection": {
				"enum": [
					"asc",
					"desc",
					"ASC",
					"DESC"
				],
				"type": "string"
			},
...
			"TxidsResponse": {
				"properties": {
					"txid": {
						"type": "string"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"idx": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"txid",
					"height",
					"idx"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "1sat-server-js",
		"contact": {}
	},
	"paths": {
...
		"/api/txos/{outpoint}": {
			"get": {
				"operationId": "GetTxoByOutpoint",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Txo"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "outpoint",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "script",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				]
			}
		},
...
		"/api/inscriptions/search": {
			"get": {
				"operationId": "GetInscriptionSearch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Txo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tag",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "dir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/SortDirection"
						}
					}
				]
			},
			"post": {
				"operationId": "PostInscriptionSearch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Txo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Inscription search. This is really powerful\n here are some really cool things you can do:\n\nSearch first-is-first: Set the sort=ASC, limit=1, offet=0",
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "tag",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "dir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/SortDirection"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TxoData"
							}
						}
					}
				}
			}
		},
		"/api/inscriptions/recent": {
			"get": {
				"operationId": "GetRecentInscriptions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Txo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/inscriptions/txid/{txid}": {
			"get": {
				"operationId": "GetInscriptionsByTxid",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Txo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "txid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/inscriptions/geohash/{geohashes}": {
			"get": {
				"operationId": "SearchGeohashes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Txo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "geohashes",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/inscriptions/{outpoint}": {
			"get": {
				"operationId": "GetTxoByOutpoint",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Txo"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "outpoint",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "script",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				]
			}
		},
		"/api/inscriptions/{outpoint}/ancestors": {
			"get": {
				"operationId": "GetAncestorsByOutpoint",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TxidsResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "outpoint",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/inscriptions/ancestors": {
			"post": {
				"operationId": "GetAncestorsByOutpoints",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TxidsResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"type": "string"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/api/inscriptions/num/{num}": {
			"get": {
				"operationId": "GetTxoByNum",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Txo"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "num",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/inscriptions/{origin}/latest": {
			"get": {
				"operationId": "GetLatestByOrigin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Txo"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "origin",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "script",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				]
			}
		},
		"/api/inscriptions/address/{address}/ancestors": {
			"get": {
				"operationId": "GetOriginTxidsByAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TxidsResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 1000,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/inscriptions/{origin}/history": {
			"get": {
				"operationId": "GetHistoryByOrigin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Txo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "origin",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 1000,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/inscriptions/{origin}/txids": {
			"get": {
				"operationId": "GetOriginTxids",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TxidsResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "origin",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/inscriptions/latest": {
			"post": {
				"operationId": "GetLatestByOrigins",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Txo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"type": "string"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}